Fix permission_required.html page
